feat(base-standalone): refactor file and construct names
