--- usecases/base-standalone/cdk.json
@@ -1,3 +1,0 @@
-{
-  "app": "npx ts-node --prefer-ts-exts bin/blea-base-sa.ts"
-}
