+++ usecases/blea-base-standalone/bin/blea-base-standalone.ts
@@ -1,5 +1,5 @@
 import * as cdk from 'aws-cdk-lib';
-import { BLEABaseSAStack } from '../lib/blea-base-sa-stack';
+import { BLEABaseStandaloneStack } from '../lib/stack/blea-base-standalone-stack';
 
 // Import parameters for each enviroment
 import { DevParameter } from '../parameter';
@@ -8,7 +8,7 @@ const app = new cdk.App();
 
 // Create stack for "Dev" environment.
 // If you have multiple environments, instantiate stacks with its parameters.
-new BLEABaseSAStack(app, 'BLEABaseSA', {
+new BLEABaseStandaloneStack(app, 'DevBLEABaseStandalone', {
   securityNotifyEmail: DevParameter.securityNotifyEmail,
   tags: { Environment: DevParameter.envName },
 });
