+++ usecases/blea-base-standalone/cdk.json
@@ -1,0 +1,3 @@
+{
+  "app": "npx ts-node --prefer-ts-exts bin/blea-base-standalone.ts"
+}
