+++ usecases/blea-base-standalone/cfn/AWS-Control-Tower-Detective-Guardrails.yaml
