+++ usecases/blea-base-standalone/lib/construct/security-detection.ts
