+++ usecases/blea-base-standalone/lib/construct/security-logging.ts
